requirements.txt
-e .
aws-cdk-lib
boto3
cdk-nag
pytest
pytest-cov
